+++ src/core/arm/arm9/bus.c
@@ -36,7 +36,7 @@ struct ARM9_MPUPerms ARM9_RegionLookup(const struct ARM946ES* ARM9, const u32 ad
     __m256i bases; memcpy(&bases, ARM9->CP15.MPURegionBase, sizeof(bases));
     addrs = _mm256_and_si256(masks, addrs);
     __m256i res = _mm256_cmpeq_epi32(addrs, bases);
-    u8 rgn = stdc_leading_zeros(_mm256_movemask_ps(_mm256_castsi256_ps(res)));
+    u8 rgn = stdc_leading_zeros((u32)_mm256_movemask_ps(_mm256_castsi256_ps(res)));
     if (rgn < 32)
         return (priv ? ARM9->CP15.MPURegionPermsPriv[31-rgn] : ARM9->CP15.MPURegionPermsUser[31-rgn]);
 #else
